Main: 1. Read CSV function to read localization-related data from CSV files and create dicts
2. Text formatting function that takes time1, time2, language and # of template and return the final text.

Removed unnescessary files/
