+++ main.py
@@ -1,67 +1,60 @@
-import datetime
-import pytz
+# coding: utf-8
+import csv
 
-timezone_source_default = 'Europe/Moscow'
-timezone_target_default = 'Asia/Shanghai'
-
-def user_date():
-    user_input = input('Please input current date as YYYY-MM-DD hh:mm: ')
-    user_input = user_input.rstrip()
-    user_input = user_input.lstrip()
-    year = user_input[0:4]
-    if int(year) > 2049 or int(year) < 2000:
-        print('Error: Years should be between 2000 and 2049')
-    else:
-        pass
-    month = user_input[5:7]
-    if int(month) > 12 or int(month) < 1:
-        print('Error: Months are between 1 and 12')
-    else:
-        pass
-    day = user_input[8:10]
-    if int(day) > 31 or int(day) < 1:
-        print('Error: Days are between 1 and 30')
-    else:
-        pass
-    hour = user_input[11:13]
-    if int(hour) > 23 or int(hour) < 0:
-        print('Error: Hours are between 0 and 23')
-    else:
-        pass
-    minute = user_input[14:16]
-    if int(minute) > 59 or int(day) < 0:
-        print('Error: Minutes are betwee 0 and 59')
-    else:
-        pass
-    date_source = (year + '-' + month + '-' + day + ' ' + hour + ':' + minute)
-    return(date_source)
-
-def convert_date_time(date_source, timezone_source, timezone_target):
-    tz1 = pytz.timezone(timezone_source)  # initial time zone
-    tz2 = pytz.timezone(timezone_target)  # target timezone
-    date_source = datetime.datetime.strptime(date_source, '%Y-%m-%d %H:%M')
-    date_source = tz1.localize(date_source)
-    date_source = date_source.astimezone(tz2)
-    year = date_source.strftime('%Y')
-    month = date_source.strftime('%m')
-    day = date_source.strftime('%d')
-    hour = date_source.strftime('%H')
-    minute = date_source.strftime('%M')
-    date_time_list = {'YYYY': year, 'MM': month, 'DD': day, 'HH': hour, 'mm': minute}
-    return date_time_list
-
-'''
-date_from_user = input('Please input current date as YYYY-MM-DD hh:mm: ')
-print(convert_date_time(date_from_user, timezone_source_default, timezone_target_default))
-'''
-
-print(convert_date_time(user_date(), timezone_source_default, timezone_target_default))
-
-months_ru_type1 = {'01': 'января', '02': 'февраля', '03': 'марта', '04': 'апреля', '05': 'мая', '06': 'июня', '07': 'июля', '08': 'августа', '09': 'сентября', '10': 'октября',  '11': 'ноября', '12': 'декабря'}
+#  function to read required info from CSV files with language related data
 
 
+def read_template(filename):
+    csv_delim = '\t'
+    # getting the first cell content and using it as ID column
+    reader = csv.reader(open(filename, 'r', encoding="utf-8"), delimiter=csv_delim)
+    for row in reader:
+        id_column_name = row[0]
+        break
+    all_rows = [] # read all rows from CSV into this list
+    reader = csv.DictReader(open(filename, 'r', encoding="utf-8"), delimiter=csv_delim)  # read CSV with TAB as delimit
+    for row in reader:
+        all_rows.append(dict(row))
+    sorted_templates = {}  # dict with all templates
+    for content in all_rows:
+        newdict_name = content[id_column_name]  # read type number from the line
+        content.pop(id_column_name)    # remove the type column
+        output = {newdict_name: content}  # create a dict with type as a key and dict as a content
+        sorted_templates = {**sorted_templates, **output}
+    return sorted_templates
 
-# template_ru_1 = (f'{DD_1} {MM_1_type1} {YYYY_1} г.')
 
-# print(template_ru_1)
+months = read_template('./localized/months_type1enru.csv')
+templates = read_template('./localized/templates.csv')
 
+
+def date_text_output(time_from_user_1, time_from_user_2, template_number, language):
+    desired_string = (templates[template_number])[language]
+    # 1 in templates stands for number in templates dict, language stands for the language in the value in the dict
+    # take month num from time_from_user, find same num in months dict, find the right language by the language key
+
+    time_formatted = {
+        'MM_1_type1': (months[time_from_user_1['month']])[language],
+        'MM_1_type2': (months[time_from_user_1['month']])[language],
+        'DD_1': time_from_user_1['day'],
+        'YYYY_1': time_from_user_1['year'],
+        'hh_1': time_from_user_1['hour'],
+        'mm_1': time_from_user_1['minute'],
+        'MM_2_type1': (months[time_from_user_2['month']])[language],
+        'MM_2_type2': (months[time_from_user_2['month']])[language],
+        'DD_2': time_from_user_2['day'],
+        'YYYY_2': time_from_user_2['year'],
+        'hh_2': time_from_user_2['hour'],
+        'mm_2': time_from_user_2['minute']
+    }
+    output = desired_string.format(**time_formatted)
+    return output
+
+
+temp_num = '10'
+time_1 = {'year': '2017', 'month': '01', 'day': '30', 'hour': '13', 'minute': '00'}
+time_2 = {'year': '2018', 'month': '02', 'day': '15', 'hour': '10', 'minute': '59'}
+lang = 'RU'
+
+
+print(date_text_output(time_1, time_2, temp_num, lang))
